[irep] replace unsigned with size_t for get_num_sub_exprs/get_sub_expr (#4074)

+++ src/irep2/templates/irep2_templates.cpp
@@ -293,8 +293,8 @@ void do_type2string<const expr2t::expr_ids>(
 template <>
 bool do_get_sub_expr<expr2tc>(
   const expr2tc &item,
-  unsigned int idx,
-  unsigned int &it,
+  size_t idx,
+  size_t &it,
   const expr2tc *&ptr)
 {
   if (idx == it)
@@ -312,8 +312,8 @@ bool do_get_sub_expr<expr2tc>(
 template <>
 bool do_get_sub_expr<std::vector<expr2tc>>(
   const std::vector<expr2tc> &item,
-  unsigned int idx,
-  unsigned int &it,
+  size_t idx,
+  size_t &it,
   const expr2tc *&ptr)
 {
   if (idx < it + item.size())
@@ -332,8 +332,8 @@ bool do_get_sub_expr<std::vector<expr2tc>>(
 template <>
 bool do_get_sub_expr_nc<expr2tc>(
   expr2tc &item,
-  unsigned int idx,
-  unsigned int &it,
+  size_t idx,
+  size_t &it,
   expr2tc *&ptr)
 {
   if (idx == it)
@@ -351,8 +351,8 @@ bool do_get_sub_expr_nc<expr2tc>(
 template <>
 bool do_get_sub_expr_nc<std::vector<expr2tc>>(
   std::vector<expr2tc> &item,
-  unsigned int idx,
-  unsigned int &it,
+  size_t idx,
+  size_t &it,
   expr2tc *&ptr)
 {
   if (idx < it + item.size())
@@ -368,13 +368,13 @@ bool do_get_sub_expr_nc<std::vector<expr2tc>>(
 }
 
 template <>
-unsigned int do_count_sub_exprs<const expr2tc>(const expr2tc &)
+size_t do_count_sub_exprs<const expr2tc>(const expr2tc &)
 {
   return 1;
 }
 
 template <>
-unsigned int
+size_t
 do_count_sub_exprs<const std::vector<expr2tc>>(const std::vector<expr2tc> &item)
 {
   return item.size();
